+++ frontend/src/components/config-steps/EncoderConfigStep.jsx
@@ -21,7 +21,7 @@ import {
   IconButton
 } from '@chakra-ui/react'
 import { InfoIcon, RepeatIcon } from '@chakra-ui/icons'
-import ParameterInput from '../buttons/ParameterInput'
+import ParameterInput from '../config-parameter-fields/ParameterInput'
 import { ODrivePropertyMappings as configurationMappings } from '../../utils/odriveUnifiedRegistry'
 import { EncoderMode } from '../../utils/odriveEnums'
