+++ frontend/src/components/config-steps/MotorConfigStep.jsx
@@ -18,7 +18,7 @@ import {
   Switch
 } from '@chakra-ui/react'
 import { InfoIcon } from '@chakra-ui/icons'
-import ParameterInput from '../buttons/ParameterInput'
+import ParameterInput from '../config-parameter-fields/ParameterInput'
 import { ODrivePropertyMappings as configurationMappings } from '../../utils/odriveUnifiedRegistry'
 
 const MotorConfigStep = ({
